First code commit
Everything works. Only thing left to do is to add functionality for a
proper database, right now the result is hard coded in database_handler.

Also, when everything works the code should be cleaned up. Seeing as
this is a quick and dirty prototype the code will not be that pretty.